Code allophones

A function to classify vowel data into contextual allophones.

code_allophones(
  .df,
  .old_col,
  .new_cols = c("allophone", "allophone_environment"),
  .pre_seg,
  .fol_seg,
  .coronals = c("T", "D", "S", "Z", "SH", "ZH", "JH", "N"),
  .voiceless = c("P", "T", "K", "CH", "F", "TH", "S", "SH")
)

Arguments

.df

The dataset containing vowel data.

.old_col

The unquoted name of the column containing the vowel labels. Often called "vowel" or "phoneme" in many datasets. Note that the function assumes Wells lexical sets (FLEECE, TRAP, etc.) rather than ARPABET (IY, AE, etc.) or IPA (i, æ, etc.). If your vowels are not already coded using Wells' labels you can quickly do so with switch_transcriptions or one of the shortcuts like arpa_to_wells

.new_cols

A vector of two strings containing the names of the columns you would like to use. By default c("allophone", "allophone_environment"). The first name becomes the name of the column containing the new allophone labels. The second column becomes the name of the column describing those labels.

.pre_seg

The unquoted name of the column that contains the labels for the previous segement. In DARLA-generated spreadsheets, this is `pre_seg` and in FastTrack-generated spreadsheets, it's `previous_sound`. Assumes ARPABET labels.

.fol_seg

The unquoted name of the column that contains the labels for the following segement. In DARLA-generated spreadsheets, this is `fol_seg` and in FastTrack-generated spreadsheets, it's `next_sound`. Assumes ARPABET labels.

.coronals

A vector of strings containing ARPABET labels for coronal consonants. By default, c("T", "D", "S", "Z", "SH", "ZH", "JH", "N"). This is used to create the `TOOT` allophone of `GOOSE`.

.voiceless

A vector of strings containing ARPABET labels for voiceless consonants. By default, c("P", "T", "K", "CH", "F", "TH", "S", "SH"). This is used to create the `PRICE` allophone of `PRIZE`.

Value

A dataframe with two additional columns. One column contains labels for the allophones and the other contains category labels for those allophones' contexts. The second column can be useful for quickly excluding certain allophones like prelaterals or prerhotics or coloring families of allophones in visualizations (such as turning all prelateral allophones gray). These two new columns are positioned immediately after the original vowel column indicated in .old_col,

Note

Here are the list of the contextual allophones that are created. Note that I largely follow my own advice about what to call elsewhere allophones, what to call prelateral allophones, and other allophones. Obviously, this list is pretty subjective and largely based on what my own research has needed, so it may not work completely for you and your research. Please contact me at joey_stanley@byu.edu if you want to see an allophone get added or if you spot an error in the coding.

FLEECE becomes

ZEAL before laterals

BEET elsewhere

KIT becomes

GUILT before laterals

NEAR before rhotics

BIG before G

BIN before M and N

BING before NG

BIT elsewhere

FACE becomes

FLAIL before laterals

VAGUE before G

X elsewhere

DRESS becomes

SHELF before laterals

SQUARE before rhotics

BEG before G

BEN before M and N

BENG before NG

BET elsewhere

TRAP becomes

TALC before laterals

BAG before G

BAN before M and N

BANG before NG

BAT elsewhere

LOT becomes

GOLF before laterals

START before rhotics

BOT elsewhere

THOUGHT becomes

FAULT before laterals

FORCE befpre rhotics

BOUGHT elsewhere

STRUT becomes

MULCH before laterals

BUT elsewhere

GOAT becomes

JOLT before laterals

BOAT elsewhere

FOOT becomes

WOLF before laterals

CURE before rhotics

PUT elsewhere

GOOSE becomes

MULE before Y

TOOT before coronals

SPOOL before laterals

BOOT elsewhere

PRICE becomes

PRICE before voiceless segments

PRIZE elsewhere

Unfortunately, it is not straightforward to customize this list but you can always copy the source code and modify the list yourself.

Alternatively, you can use forcats::fct_collapse() to collapse distinctions that you don't need. See example code below.

You can also of course create your own allophones if desired. Note that some allophones depend on other environmental information like syllable structure and morpheme/word boundaries, or they may be entirely lexical (FORCE vs. NORTH). They may be more complicated than what ARPABET can code for (MARY, MERRY, and MARRY) or just inconsistently coded. For the sake of simplicity, these allophones are not included in this function.

The environments therefore are the following

"prelateral" includes ZEAL, GUILT, FLAIL, SHELF, TALC, GOLF, FAULT, MULCH, JOLT, WOLF, SPOOL

"prerhotic" includes NEAR, SQUARE, START, FORCE, CURE

"prevelar" includes BIG, VAGUE, BEG, BAG,

"prenasal" includes BIN, BEN, BAN

"prevelarnasal" includes BING, BENG, BANG

"prevoiceless" includes PRICE

"post-Y" includes MULE

"postcoronal" includes TOOT

"elsewhere" includes BEET, BIT, BAIT, BET, BAT, BOT, BOUGHT, BUT, BOAT, PUT, BOOT, PRIZE

Examples

suppressPackageStartupMessages(library(tidyverse))
#> Warning: package ‘ggplot2’ was built under R version 4.1.2
#> Warning: package ‘tibble’ was built under R version 4.1.2
#> Warning: package ‘tidyr’ was built under R version 4.1.2
#> Warning: package ‘dplyr’ was built under R version 4.1.2

# Get some sample DARLA data to play with
darla <- joeysvowels::darla %>%
  select(word, vowel, pre_seg, fol_seg) %>%
  mutate(phoneme = joeyr:::arpa_to_wells(vowel), .after = vowel)

# Basic usage
darla %>%
  code_allophones(.old_col = phoneme, .fol_seg = fol_seg, .pre_seg = pre_seg) %>%
  slice_sample(n = 20)
#>         word vowel phoneme allophone allophone_environment pre_seg fol_seg
#> 1    DISEASE    IY  FLEECE      BEET             elsewhere       Z       Z
#> 2      FAULT    AO THOUGHT     FAULT            prelateral       F       L
#> 3  NECESSARY    EH   DRESS    SQUARE             prerhotic       S       R
#> 4  BILLBOARD    IH     KIT     GUILT            prelateral       B       L
#> 5         TO    UW   GOOSE      TOOT           postcoronal       T      DH
#> 6  VALENTINE    AE    TRAP      TALC            prelateral       V       L
#> 7        ALL    AO THOUGHT     FAULT            prelateral       D       L
#> 8       LIKE    AY   PRICE     PRICE          prevoiceless       L       K
#> 9    FLOWING    IH     KIT      BING         prevelarnasal     OW1      NG
#> 10       MAY    EY    FACE      BAIT             elsewhere       M       N
#> 11     SPACE    EY    FACE      BAIT             elsewhere       P       S
#> 12     POINT    OY  CHOICE    CHOICE             elsewhere       P       N
#> 13     THEFT    EH   DRESS       BET             elsewhere      TH       F
#> 14 SOMETHING    AH   STRUT       BUT             elsewhere       S       M
#> 15        BY    AY   PRICE     PRIZE             elsewhere       B      DH
#> 16  PATIENTS    EY    FACE      BAIT             elsewhere       P      SH
#> 17       ARE    ER   NURSE     NURSE             elsewhere       Z        
#> 18       HOT    AA     LOT       BOT             elsewhere      HH       T
#> 19        TO    UW   GOOSE      TOOT           postcoronal       T       Y
#> 20     AFTER    AE    TRAP       BAT             elsewhere       N       F

# Specify the names of the new columns with the `.new_cols` argument
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  slice_sample(n = 20)
#>         word vowel phoneme allophone  environment pre_seg fol_seg
#> 1    ENTERED    EH   DRESS       BEN     prenasal       N       N
#> 2       BLUE    UW   GOOSE      BOOT    elsewhere       L       T
#> 3     CHARGE    AA     LOT     START    prerhotic      CH       R
#> 4    READING    IY  FLEECE      BEET    elsewhere       R       D
#> 5  FRAMEWORK    EY    FACE      BAIT    elsewhere       R       M
#> 6     HYBRID    AY   PRICE     PRIZE    elsewhere      HH       B
#> 7         TO    UW   GOOSE      TOOT  postcoronal       T       F
#> 8      CHILI    IY  FLEECE      BEET    elsewhere       L       P
#> 9      PLACE    EY    FACE      BAIT    elsewhere       L       S
#> 10   UPRIGHT    AY   PRICE     PRICE prevoiceless       R       T
#> 11        DO    UW   GOOSE      TOOT  postcoronal       D      DH
#> 12      NEWT    UW   GOOSE      TOOT  postcoronal       N       T
#> 13        SO    OW    GOAT      BOAT    elsewhere       S     AH0
#> 14      SAME    EY    FACE      BAIT    elsewhere       S       M
#> 15      TIME    AY   PRICE     PRIZE    elsewhere       T       M
#> 16   NUCLEAR    UW   GOOSE      TOOT  postcoronal       N       K
#> 17       FAR    AA     LOT     START    prerhotic       F       R
#> 18    MEXICO    OW    GOAT      BOAT    elsewhere       K        
#> 19   HOSTESS    OW    GOAT      BOAT    elsewhere      HH       S
#> 20 REQUIRING    AY   PRICE     PRIZE    elsewhere       W       R

# Filtering by environment is straightforward
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  filter(environment == "elsewhere") %>%
  slice_sample(n = 20)
#>           word vowel phoneme allophone environment pre_seg fol_seg
#> 1         POLO    OW    GOAT      BOAT   elsewhere       L        
#> 2        HEAVY    IY  FLEECE      BEET   elsewhere       V       T
#> 3      FACTORY    AE    TRAP       BAT   elsewhere       F       K
#> 4         POST    OW    GOAT      BOAT   elsewhere       P       S
#> 5        ABOUT    AH   STRUT       BUT   elsewhere       S       B
#> 6      SPEAKER    IY  FLEECE      BEET   elsewhere       P       K
#> 7    GRADUATES    AE    TRAP       BAT   elsewhere       R      JH
#> 8       VELVET    AH   STRUT       BUT   elsewhere       V       T
#> 9           US    EH   DRESS       BET   elsewhere     UW1       S
#> 10     KEEPING    IY  FLEECE      BEET   elsewhere       K       P
#> 11       COULD    UH    FOOT       PUT   elsewhere       K       D
#> 12       CABIN    AE    TRAP       BAT   elsewhere       K       B
#> 13      MOMENT    AH   STRUT       BUT   elsewhere       M       N
#> 14      GROUND    AW   MOUTH     MOUTH   elsewhere       R       N
#> 15       FRAME    EY    FACE      BAIT   elsewhere       R       M
#> 16        TIME    AY   PRICE     PRIZE   elsewhere       T       M
#> 17    THIRTEEN    IY  FLEECE      BEET   elsewhere       T       N
#> 18     SUCCESS    EH   DRESS       BET   elsewhere       S       S
#> 19 LIMITATIONS    IH     KIT       BIT   elsewhere       M       T
#> 20     TRAINED    EY    FACE      BAIT   elsewhere       R       N
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  filter(!environment %in% c("prerhotic", "prevelarnasal", "prevelar")) %>%
  slice_sample(n = 20)
#>         word vowel phoneme allophone environment pre_seg fol_seg
#> 1    PROGRAM    OW    GOAT      BOAT   elsewhere       R       G
#> 2    REACHED    IY  FLEECE      BEET   elsewhere       R      CH
#> 3        ALL    AO THOUGHT     FAULT  prelateral       T       L
#> 4        BUY    AY   PRICE     PRIZE   elsewhere       B        
#> 5      WORDS    ER   NURSE     NURSE   elsewhere       W       D
#> 6       EACH    IY  FLEECE      BEET   elsewhere       L      CH
#> 7   ASTEROID    AE    TRAP       BAT   elsewhere     IY0       S
#> 8    CHINESE    IY  FLEECE      BEET   elsewhere       N       Z
#> 9   COTTAGES    AA     LOT       BOT   elsewhere       K       T
#> 10     FIRST    ER   NURSE     NURSE   elsewhere       F       S
#> 11   PLEASES    IY  FLEECE      BEET   elsewhere       L       Z
#> 12        TO    UW   GOOSE      TOOT postcoronal       T       G
#> 13    WORKED    ER   NURSE     NURSE   elsewhere       W       K
#> 14    AROUND    ER   NURSE     NURSE   elsewhere      NG     AW1
#> 15     THREE    IY  FLEECE      BEET   elsewhere       R       K
#> 16 EDUCATION    EY    FACE      BAIT   elsewhere       K      SH
#> 17   REFRESH    EH   DRESS       BET   elsewhere       R      SH
#> 18     PLACE    EY    FACE      BAIT   elsewhere       L       S
#> 19       MIA    IY  FLEECE      BEET   elsewhere       M     AH0
#> 20       YEN    EH   DRESS       BEN    prenasal       Y       N

# Some users may want to supply their own list of coronal consonants.
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg,
                  .coronals = c("T", "D", "S", "Z", "SH", "ZH", "JH", "N", "Y")) %>%
  filter(phoneme == "GOOSE") %>%
  slice_sample(n = 20)
#>        word vowel phoneme allophone environment pre_seg fol_seg
#> 1  PRODUCES    UW   GOOSE      TOOT postcoronal       D       S
#> 2     GROOM    UW   GOOSE      BOOT   elsewhere       R       M
#> 3      HUGE    UW   GOOSE      MULE      post-Y       Y      JH
#> 4        TO    UW   GOOSE      TOOT postcoronal       T       Y
#> 5        TO    UW   GOOSE      TOOT postcoronal       T      HH
#> 6   FOOLISH    UW   GOOSE     SPOOL  prelateral       F       L
#> 7      INTO    UW   GOOSE      TOOT postcoronal       T       P
#> 8       LOU    UW   GOOSE      BOOT   elsewhere       L        
#> 9        TO    UW   GOOSE      TOOT postcoronal       T       K
#> 10      USE    UW   GOOSE      MULE      post-Y       Y       Z
#> 11     SOON    UW   GOOSE      TOOT postcoronal       S       N
#> 12      FEW    UW   GOOSE      MULE      post-Y       Y       S
#> 13      TWO    UW   GOOSE      TOOT postcoronal       T        
#> 14       TO    UW   GOOSE      TOOT postcoronal       T     IY0
#> 15  BALLOON    UW   GOOSE      BOOT   elsewhere       L       N
#> 16     TOOL    UW   GOOSE      TOOT postcoronal       T       L
#> 17  THROUGH    UW   GOOSE      BOOT   elsewhere       R      HH
#> 18    LOOPS    UW   GOOSE      BOOT   elsewhere       L       P
#> 19 SCHOOL'S    UW   GOOSE     SPOOL  prelateral       K       L
#> 20       TO    UW   GOOSE      TOOT postcoronal       T     AH0

# Other users may want to specify their own list of voiceless consonants. 
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg,
                  .voiceless = c("P", "T", "K", "CH", "F", "TH", "S", "SH", "X")) %>%
  filter(phoneme == "PRICE") %>%
  slice_sample(n = 20)
#>         word vowel phoneme allophone  environment pre_seg fol_seg
#> 1   ASSIGNED    AY   PRICE     PRIZE    elsewhere       S       N
#> 2       HIGH    AY   PRICE     PRIZE    elsewhere      HH       M
#> 3      BRIAN    AY   PRICE     PRIZE    elsewhere       R     AH0
#> 4      TIGER    AY   PRICE     PRIZE    elsewhere       T       G
#> 5       DIET    AY   PRICE     PRIZE    elsewhere       D     AH0
#> 6       LIKE    AY   PRICE     PRICE prevoiceless       L       K
#> 7      QUITE    AY   PRICE     PRICE prevoiceless       W       T
#> 8   SUPPLIES    AY   PRICE     PRIZE    elsewhere       L       Z
#> 9  VARIETIES    AY   PRICE     PRIZE    elsewhere     ER0     AH0
#> 10     PRIME    AY   PRICE     PRIZE    elsewhere       R       M
#> 11      LINE    AY   PRICE     PRIZE    elsewhere       L       N
#> 12     ITEMS    AY   PRICE     PRICE prevoiceless               T
#> 13    DESIGN    AY   PRICE     PRIZE    elsewhere       Z       N
#> 14      LIFE    AY   PRICE     PRICE prevoiceless       L       F
#> 15    SLIDES    AY   PRICE     PRIZE    elsewhere       L       D
#> 16    HYBRID    AY   PRICE     PRIZE    elsewhere      HH       B
#> 17       SLY    AY   PRICE     PRIZE    elsewhere       L     AH0
#> 18     DRIVE    AY   PRICE     PRIZE    elsewhere       R       V
#> 19      HIGH    AY   PRICE     PRICE prevoiceless      HH       S
#> 20      EYES    AY   PRICE     PRIZE    elsewhere       N       Z

# Collapsing distinctions can be done post hoc (though it may take extra work to get the environment column to match.)
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  # Get a subset for demonstration purposes
  filter(allophone %in% c("BIT", "BIG")) %>%
  group_by(allophone) %>%
  slice_sample(n = 5) %>%
  ungroup() %>%
  # Now collapse distinctions
  mutate(allophone = fct_collapse(allophone, "BIT" = c("BIT", "BIG")),
         environment = ifelse(allophone == "BIT", "elsewhere", allophone))
#> # A tibble: 10 × 7
#>    word      vowel phoneme allophone environment pre_seg fol_seg
#>    <chr>     <chr> <fct>   <fct>     <chr>       <chr>   <chr>  
#>  1 BEGINNING IH    KIT     BIT       elsewhere   B       G      
#>  2 BIG       IH    KIT     BIT       elsewhere   B       G      
#>  3 EXISTENCE IH    KIT     BIT       elsewhere   IY0     G      
#>  4 EXOTIC    IH    KIT     BIT       elsewhere   V       G      
#>  5 TRIGGER   IH    KIT     BIT       elsewhere   R       G      
#>  6 DID       IH    KIT     BIT       elsewhere   D       D      
#>  7 WITH      IH    KIT     BIT       elsewhere   W       TH     
#>  8 HIS       IH    KIT     BIT       elsewhere   HH      Z      
#>  9 LISTEN    IH    KIT     BIT       elsewhere   L       S      
#> 10 SICK      IH    KIT     BIT       elsewhere   S       K      

# Creating new allophones depends on the complexity of the allophone         
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  # Create voice and voiceless distinctions for MOUTH
  mutate(allophone = case_when(phoneme == "MOUTH" & fol_seg %in% c("P", "T", "K", "CH", "F", "TH", "S", "SH") ~ "BOUT",
                               phoneme == "MOUTH" ~ "LOUD",
                               TRUE ~ allophone),
         environment = if_else(allophone == "BOUT",  "prevoiceless", environment)) %>%
  # Get a subset for demonstration purposes
  filter(phoneme == "MOUTH") %>%
  group_by(allophone) %>%
  slice_sample(n = 5) %>%
  ungroup()
#> # A tibble: 10 × 7
#>    word         vowel phoneme allophone environment  pre_seg fol_seg
#>    <chr>        <chr> <fct>   <chr>     <chr>        <chr>   <chr>  
#>  1 ABOUT        AW    MOUTH   BOUT      prevoiceless B       T      
#>  2 ABOUT        AW    MOUTH   BOUT      prevoiceless B       T      
#>  3 ABOUT        AW    MOUTH   BOUT      prevoiceless B       T      
#>  4 HOUSE        AW    MOUTH   BOUT      prevoiceless HH      S      
#>  5 OUT          AW    MOUTH   BOUT      prevoiceless T       T      
#>  6 THOUSANDS    AW    MOUTH   LOUD      elsewhere    TH      Z      
#>  7 AROUND       AW    MOUTH   LOUD      elsewhere    ER0     N      
#>  8 ANNOUNCEMENT AW    MOUTH   LOUD      elsewhere    N       N      
#>  9 DOWN         AW    MOUTH   LOUD      elsewhere    D       N      
#> 10 COUNTY       AW    MOUTH   LOUD      elsewhere    K       N