ANAE Vowel Normalization

This is a a tidyverse-compatible function that makes it easy to normalize your data using the method described in the Atlas of North American English. NOTE: This is brand new and has not been tested robustly.

norm_anae(df, hz_cols, vowel_id, speaker_id, use_telsur_g = TRUE)

Arguments

df
The dataframe containing the formant measurements you want to normalize.

hz_cols
A list of columns (unquoted) containing the formant measurements themselves.

vowel_id
The name of the column containing unique identifiers per vowel.

speaker_id
The name of the column containing unique identifiers per speaker (usually the column containing the speaker name).

use_telsur_g
By default, this will use the Telsur G value (6.896874) listed in the ANAE. If set to FALSE, it will calculate the G value based on the dataset.

Value

The same dataframe, but with new column(s), suffixed with "_anae" that have the normalized data.

Details

Note that the data must be grouped by speaker prior to running the function.

It is unclear how the ANAE function should work with trajectory data. This function pools all data together and normalizes it together, which means one small modification was required to calculate the G value: I had to add the average number of time points per vowel token in the denominator. Not sure if that's how it should be done, but it makes sense to me and returns sensible results.

Examples

library(tidyverse)
df <- joeysvowels::idahoans

# Run the function. Since this data doesn't have a column for token ids, I'll create one here.
df %>%
   rowid_to_column("token_id") %>%
   group_by(speaker) %>%
   norm_anae(hz_cols = c(F1, F2), vowel_id = token_id, speaker_id = speaker) %>%
   ungroup() %>%
   select(F1, F2, F1_anae, F2_anae) # <- just the relevant columns

#> # A tibble: 1,100 x 4
#>       F1    F2 F1_anae F2_anae
#>    <dbl> <dbl>   <dbl>   <dbl>
#>  1  699. 1655.    714.   1690.
#>  2  685. 1360.    700.   1388.
#>  3  713. 1507.    728.   1539.
#>  4  801. 1143.    818.   1167.
#>  5  757. 1258.    772.   1284.
#>  6  804. 1403.    821.   1432.
#>  7  664. 1279.    678.   1306.
#>  8  757. 1325.    773.   1353.
#>  9  730. 1578.    746.   1611.
#> 10  700. 1546.    715.   1578.
#> # … with 1,090 more rows