∆F Normalization

Normalize vowel formant measurements with ∆F (see Johnson 2020). This function is intended to be used within a tidyverse pipeline.

norm_deltaF(df, .F1, .F2, .F3, .F4)

Arguments

df
The data frame containing the formant measurements you want to normalize

.F1, .F2, .F3, .F4
The (unquoted) name of the column containing the F1 measurements. The first three are required, but you may leave off F4.

Value

The original dataframe with new columns (suffixed with _deltaF) containing the normalized measurements.

Details

You will need to group the data by speaker with group_by() before applying this function if you want to normalize the data by speaker.

The data must be numeric, and there cannot be any NAs. So, if you're using data extracted from Praat, you may have to filter out bad F3 and F4 data and then convert the column to numeric. See the example.

Note that this is a new function and has not been tested robustly yet.

References

Johnson, Keith. The ΔF Method of Vocal Tract Length Normalization for Vowels. Laboratory Phonology: Journal of the Association for Laboratory Phonology 11, no. 1 (July 22, 2020): 10. https://doi.org/10.5334/labphon.196.

Examples

library(tidyverse)
df <- joeysvowels::idahoans

 # Run the function
 df %>%
   group_by(speaker) %>%
   norm_deltaF(F1, F2, F3)

#> # A tibble: 1,100 x 10
#> # Groups:   speaker [10]
#>    speaker sex    vowel    F1    F2    F3 F1_norm F2_norm F3_norm    F4
#>    <fct>   <chr>  <chr> <dbl> <dbl> <dbl>   <dbl>   <dbl>   <dbl> <dbl>
#>  1 01      female AA     699. 1655. 2019.   0.626    1.48    1.81 3801.
#>  2 01      female AA     685. 1360. 1914.   0.614    1.22    1.71 4257.
#>  3 01      female AA     713. 1507. 2460.   0.639    1.35    2.20 3617.
#>  4 01      female AA     801. 1143. 1868.   0.718    1.02    1.67 2908.
#>  5 01      female AA     757. 1258. 1772.   0.677    1.13    1.59 2778.
#>  6 01      female AA     804. 1403. 2339.   0.720    1.26    2.09 4299.
#>  7 01      female AA     664. 1279. 1714.   0.595    1.15    1.53 2103.
#>  8 01      female AA     757. 1325. 1929.   0.678    1.19    1.73 2660.
#>  9 01      female AA     730. 1578. 2297.   0.654    1.41    2.06 2963.
#> 10 01      female AA     700. 1546. 2109.   0.627    1.38    1.89 3432.
#> # … with 1,090 more rows