spread() on multiple values.

A modification of tidyr::spread that allows you to use it with multiple `value` columns.

spread_n(df, key, ...)

Arguments

df
The dataframe you want to modify.

key
The column you want to have spread across multiple columns. For tall vowel data, this is usually the "formant" column.

...
The data you want to have fill those new columns.

Details

All credit for this function goes to Dan Sullivan on the RStudio Forums who answered my quesiton here. Apparently Kieran Healey seemed to like it and wrote a blog post about the function here and does a really good job at explaning when you might need to use a function like this. A while later, when Hadley Wickham asked for interesting cases where tidy eval is used, I commented with this function. He replied with a shorter version, which is what is implemented now.

I still don't fully understand the code, but it works exactly how I want it to so I keep using it.

Examples

require(tidyr)

# Example 1
df1 <- data.frame(word = rep(c("dog", "cat", "fish"), 2),
                  formant = rep(c("F1", "F2"), each = 3),
                  hz = c(490, 500, 510, 1590, 1600, 1610),
                  bark = bark(c(490, 500, 510, 1590, 1600, 1610)))

#> Error in bark(c(490, 500, 510, 1590, 1600, 1610)): could not find function "bark"

# Now spread them with each formant and type of measurement in its own column.
df1 %>%
  spread_n(formant, c(hz, bark))

#> Error in eval(lhs, parent, parent): object 'df1' not found

# Example 2
df2 <- data.frame(speaker = rep(c("Amy", "Bob"), each = 3),
                  word = rep(c("dog", "cat", "fish"), 2),
                  F1 = c(490, 500, 510, 390, 400, 410),
                  F2 = c(1590, 1600, 1610, 1390, 1400, 1410))
# Change it to one row per speaker
df2 %>%
  spread_n(word, c(F1, F2))

#> Error in spread_n(., word, c(F1, F2)): could not find function "spread_n"

# Change it to one row per word
df2 %>%
  spread_n(speaker, c(F1, F2))

#> Error in spread_n(., speaker, c(F1, F2)): could not find function "spread_n"