Code allophones

A function to classify vowel data into contextual allophones.

code_allophones(
  .df,
  .old_col,
  .new_cols = c("allophone", "allophone_environment"),
  .pre_seg,
  .fol_seg,
  .coronals = c("T", "D", "S", "Z", "SH", "ZH", "JH", "N"),
  .voiceless = c("P", "T", "K", "CH", "F", "TH", "S", "SH")
)

Arguments

.df

The dataset containing vowel data.

.old_col

The unquoted name of the column containing the vowel labels. Often called "vowel" or "phoneme" in many datasets. Note that the function assumes Wells lexical sets (FLEECE, TRAP, etc.) rather than ARPABET (IY, AE, etc.) or IPA (i, æ, etc.). If your vowels are not already coded using Wells' labels you can quickly do so with switch_transcriptions or one of the shortcuts like arpa_to_wells

.new_cols

A vector of two strings containing the names of the columns you would like to use. By default c("allophone", "allophone_environment"). The first name becomes the name of the column containing the new allophone labels. The second column becomes the name of the column describing those labels.

.pre_seg

The unquoted name of the column that contains the labels for the previous segement. In DARLA-generated spreadsheets, this is `pre_seg` and in FastTrack-generated spreadsheets, it's `previous_sound`. Assumes ARPABET labels.

.fol_seg

The unquoted name of the column that contains the labels for the following segement. In DARLA-generated spreadsheets, this is `fol_seg` and in FastTrack-generated spreadsheets, it's `next_sound`. Assumes ARPABET labels.

.coronals

A vector of strings containing ARPABET labels for coronal consonants. By default, c("T", "D", "S", "Z", "SH", "ZH", "JH", "N"). This is used to create the `TOOT` allophone of `GOOSE`.

.voiceless

A vector of strings containing ARPABET labels for voiceless consonants. By default, c("P", "T", "K", "CH", "F", "TH", "S", "SH"). This is used to create the `PRICE` allophone of `PRIZE`.

Value

A dataframe with two additional columns. One column contains labels for the allophones and the other contains category labels for those allophones' contexts. The second column can be useful for quickly excluding certain allophones like prelaterals or prerhotics or coloring families of allophones in visualizations (such as turning all prelateral allophones gray). These two new columns are positioned immediately after the original vowel column indicated in .old_col,

Note

Here are the list of the contextual allophones that are created. Note that I largely follow my own advice about what to call elsewhere allophones, what to call prelateral allophones, and other allophones. Obviously, this list is pretty subjective and largely based on what my own research has needed, so it may not work completely for you and your research. Please contact me at joey_stanley@byu.edu if you want to see an allophone get added or if you spot an error in the coding.

FLEECE becomes

ZEAL before laterals

BEET elsewhere

KIT becomes

GUILT before laterals

NEAR before rhotics

BIG before G

BIN before M and N

BING before NG

BIT elsewhere

FACE becomes

FLAIL before laterals

VAGUE before G

X elsewhere

DRESS becomes

SHELF before laterals

SQUARE before rhotics

BEG before G

BEN before M and N

BENG before NG

BET elsewhere

TRAP becomes

TALC before laterals

BAG before G

BAN before M and N

BANG before NG

BAT elsewhere

LOT becomes

GOLF before laterals

START before rhotics

BOT elsewhere

THOUGHT becomes

FAULT before laterals

FORCE befpre rhotics

BOUGHT elsewhere

STRUT becomes

MULCH before laterals

BUT elsewhere

GOAT becomes

JOLT before laterals

BOAT elsewhere

FOOT becomes

WOLF before laterals

CURE before rhotics

PUT elsewhere

GOOSE becomes

MULE before Y

TOOT before coronals

SPOOL before laterals

BOOT elsewhere

PRICE becomes

PRICE before voiceless segments

PRIZE elsewhere

Unfortunately, it is not straightforward to customize this list but you can always copy the source code and modify the list yourself.

Alternatively, you can use forcats::fct_collapse() to collapse distinctions that you don't need. See example code below.

You can also of course create your own allophones if desired. Note that some allophones depend on other environmental information like syllable structure and morpheme/word boundaries, or they may be entirely lexical (FORCE vs. NORTH). They may be more complicated than what ARPABET can code for (MARY, MERRY, and MARRY) or just inconsistently coded. For the sake of simplicity, these allophones are not included in this function.

The environments therefore are the following

"prelateral" includes ZEAL, GUILT, FLAIL, SHELF, TALC, GOLF, FAULT, MULCH, JOLT, WOLF, SPOOL

"prerhotic" includes NEAR, SQUARE, START, FORCE, CURE

"prevelar" includes BIG, VAGUE, BEG, BAG,

"prenasal" includes BIN, BEN, BAN

"prevelarnasal" includes BING, BENG, BANG

"prevoiceless" includes PRICE

"post-Y" includes MULE

"postcoronal" includes TOOT

"elsewhere" includes BEET, BIT, BAIT, BET, BAT, BOT, BOUGHT, BUT, BOAT, PUT, BOOT, PRIZE

Examples

suppressPackageStartupMessages(library(tidyverse))

# Get some sample DARLA data to play with
darla <- joeysvowels::darla %>%
  select(word, vowel, pre_seg, fol_seg) %>%
  mutate(phoneme = joeyr:::arpa_to_wells(vowel), .after = vowel)

# Basic usage
darla %>%
  code_allophones(.old_col = phoneme, .fol_seg = fol_seg, .pre_seg = pre_seg) %>%
  slice_sample(n = 20)
#>         word vowel phoneme allophone allophone_environment pre_seg fol_seg
#> 1   MILITARY    IH     KIT       BIT             elsewhere       L       T
#> 2    HAPPILY    AE    TRAP       BAT             elsewhere      HH       P
#> 3    SYSTEMS    AH   STRUT       BUT             elsewhere       T       M
#> 4       RAIN    EY    FACE      BAIT             elsewhere       R       N
#> 5        TWO    UW   GOOSE      TOOT           postcoronal       T       R
#> 6   TEACHERS    ER   NURSE     NURSE             elsewhere      CH       Z
#> 7         ON    AO THOUGHT    BOUGHT             elsewhere       T       N
#> 8  DIFFICULT    IH     KIT       BIT             elsewhere       D       F
#> 9    HOWEVER    ER   NURSE     NURSE             elsewhere       V     IH1
#> 10   SCHOOLS    UW   GOOSE     SPOOL            prelateral       K       L
#> 11   WINDOWS    OW    GOAT      BOAT             elsewhere       D       Z
#> 12   VIEWERS    ER   NURSE     NURSE             elsewhere     UW1       Z
#> 13     PATHS    AE    TRAP       BAT             elsewhere       P      TH
#> 14    SIMPLE    IH     KIT       BIN              prenasal       S       M
#> 15        BY    AY   PRICE     PRIZE             elsewhere       B      DH
#> 16        TO    UW   GOOSE      TOOT           postcoronal       T       Y
#> 17    FLOWER    AW   MOUTH     MOUTH             elsewhere       L     ER0
#> 18      YOUR    UH    FOOT      CURE             prerhotic       Y       R
#> 19 FINANCIAL    AE    TRAP       BAN              prenasal       N       N
#> 20      GOOD    UH    FOOT       PUT             elsewhere       G       D

# Specify the names of the new columns with the `.new_cols` argument
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  slice_sample(n = 20)
#>              word vowel phoneme allophone   environment pre_seg fol_seg
#> 1           SMALL    AO THOUGHT     FAULT    prelateral       M       L
#> 2             RUN    AH   STRUT       BUT     elsewhere       R       N
#> 3              ON    AO THOUGHT    BOUGHT     elsewhere       T       N
#> 4           SWEAT    EH   DRESS       BET     elsewhere       W       T
#> 5           EARLY    IY  FLEECE      BEET     elsewhere       L      DH
#> 6            LOUD    AW   MOUTH     MOUTH     elsewhere       L       D
#> 7    NEIGHBORHOOD    EY    FACE      BAIT     elsewhere       N       B
#> 8              ON    AO THOUGHT    BOUGHT     elsewhere       V       N
#> 9         PLEASES    IH     KIT       BIT     elsewhere       Z       Z
#> 10         BEHAVE    EY    FACE      BAIT     elsewhere      HH       V
#> 11     IDENTIFIED    AY   PRICE     PRIZE     elsewhere       F       D
#> 12           FEEL    IY  FLEECE      ZEAL    prelateral       F       L
#> 13      PREPARING    IH     KIT      BING prevelarnasal       R      NG
#> 14           HALF    AE    TRAP       BAT     elsewhere      HH       F
#> 15       DISTRICT    IH     KIT       BIT     elsewhere       D       S
#> 16        REGULAR    EY    FACE     VAGUE      prevelar       R       G
#> 17     ASSESSMENT    AH   STRUT       BUT     elsewhere       M       N
#> 18 IMPLEMENTATION    EY    FACE      BAIT     elsewhere       T      SH
#> 19        WILLING    IH     KIT      BING prevelarnasal       L      NG
#> 20        QUARTER    AO THOUGHT     FORCE     prerhotic       W       R

# Filtering by environment is straightforward
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  filter(environment == "elsewhere") %>%
  slice_sample(n = 20)
#>          word vowel phoneme allophone environment pre_seg fol_seg
#> 1        LOGS    AO THOUGHT    BOUGHT   elsewhere       L       G
#> 2       ABOUT    AW   MOUTH     MOUTH   elsewhere       B       T
#> 3    SLEEPING    IY  FLEECE      BEET   elsewhere       L       P
#> 4         OUT    AW   MOUTH     MOUTH   elsewhere       T       T
#> 5      CLOSER    OW    GOAT      BOAT   elsewhere       L       Z
#> 6     CHAMBER    ER   NURSE     NURSE   elsewhere       B     AE1
#> 7       LINDA    AH   STRUT       BUT   elsewhere       D       N
#> 8      GRADES    EY    FACE      BAIT   elsewhere       R       D
#> 9  BOUNDARIES    AW   MOUTH     MOUTH   elsewhere       B       N
#> 10     TREATS    IY  FLEECE      BEET   elsewhere       R       T
#> 11      THROW    OW    GOAT      BOAT   elsewhere       R     IH1
#> 12        RUN    AH   STRUT       BUT   elsewhere       R       N
#> 13    POCKETS    AA     LOT       BOT   elsewhere       P       K
#> 14       ABLE    EY    FACE      BAIT   elsewhere       Z       B
#> 15  JEFFERSON    AH   STRUT       BUT   elsewhere       S       N
#> 16  MARIJUANA    AH   STRUT       BUT   elsewhere       N        
#> 17 WASHINGTON    AO THOUGHT    BOUGHT   elsewhere       W      SH
#> 18 INNOVATION    EY    FACE      BAIT   elsewhere       V      SH
#> 19         ON    AO THOUGHT    BOUGHT   elsewhere       D       N
#> 20      LATER    EY    FACE      BAIT   elsewhere       L       T
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  filter(!environment %in% c("prerhotic", "prevelarnasal", "prevelar")) %>%
  slice_sample(n = 20)
#>         word vowel phoneme allophone environment pre_seg fol_seg
#> 1  AFTERNOON    AE    TRAP       BAT   elsewhere       Z       F
#> 2      DEATH    EH   DRESS       BET   elsewhere       D      TH
#> 3        HAD    AE    TRAP       BAT   elsewhere      HH       D
#> 4      THOSE    OW    GOAT      BOAT   elsewhere      DH       Z
#> 5    MEMBERS    EH   DRESS       BEN    prenasal       M       M
#> 6         OR    ER   NURSE     NURSE   elsewhere     IY0     EY1
#> 7       LENA    IY  FLEECE      BEET   elsewhere       L       N
#> 8    YOUNGER    ER   NURSE     NURSE   elsewhere       G      DH
#> 9       LONG    AO THOUGHT    BOUGHT   elsewhere       L      NG
#> 10     WORDS    ER   NURSE     NURSE   elsewhere       W       D
#> 11  KITCHENS    IH     KIT       BIT   elsewhere       K      CH
#> 12       ALL    AO THOUGHT     FAULT  prelateral       D       L
#> 13        ON    AO THOUGHT    BOUGHT   elsewhere       T       N
#> 14 FEROCIOUS    OW    GOAT      BOAT   elsewhere     ER0      SH
#> 15   SUCCESS    EH   DRESS       BET   elsewhere       S       S
#> 16      GAME    EY    FACE      BAIT   elsewhere       G       M
#> 17  AILMENTS    EY    FACE     FLAIL  prelateral               L
#> 18   FOLIAGE    IH     KIT       BIT   elsewhere     IY0      JH
#> 19    EXPECT    EH   DRESS       BET   elsewhere       P       K
#> 20     SEEDS    IY  FLEECE      BEET   elsewhere       S       D

# Some users may want to supply their own list of coronal consonants.
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg,
                  .coronals = c("T", "D", "S", "Z", "SH", "ZH", "JH", "N", "Y")) %>%
  filter(phoneme == "GOOSE") %>%
  slice_sample(n = 20)
#>           word vowel phoneme allophone environment pre_seg fol_seg
#> 1         DULY    UW   GOOSE      TOOT postcoronal       D       L
#> 2          TOO    UW   GOOSE      TOOT postcoronal       T       K
#> 3    MOLECULES    UW   GOOSE      MULE      post-Y       Y       L
#> 4        DOING    UW   GOOSE      TOOT postcoronal       D     IH0
#> 5           TO    UW   GOOSE      TOOT postcoronal       T        
#> 6           TO    UW   GOOSE      TOOT postcoronal       T     IH0
#> 7         INTO    UW   GOOSE      TOOT postcoronal       T     SIL
#> 8  CONTRIBUTED    UW   GOOSE      MULE      post-Y       Y       T
#> 9   INTRODUCES    UW   GOOSE      TOOT postcoronal       D       S
#> 10        POOL    UW   GOOSE     SPOOL  prelateral       P       L
#> 11     FOOLISH    UW   GOOSE     SPOOL  prelateral       F       L
#> 12        INTO    UW   GOOSE      TOOT postcoronal       T       W
#> 13     SCHOOLS    UW   GOOSE     SPOOL  prelateral       K       L
#> 14      UNIQUE    UW   GOOSE      MULE      post-Y       Y       N
#> 15        POOL    UW   GOOSE     SPOOL  prelateral       P       L
#> 16        POOR    UW   GOOSE      BOOT   elsewhere       P       R
#> 17     PREVIEW    UW   GOOSE      MULE      post-Y       Y      HH
#> 18        INTO    UW   GOOSE      TOOT postcoronal       T     AH0
#> 19          TO    UW   GOOSE      TOOT postcoronal       T      HH
#> 20      HUMANS    UW   GOOSE      MULE      post-Y       Y       M

# Other users may want to specify their own list of voiceless consonants. 
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg,
                  .voiceless = c("P", "T", "K", "CH", "F", "TH", "S", "SH", "X")) %>%
  filter(phoneme == "PRICE") %>%
  slice_sample(n = 20)
#>          word vowel phoneme allophone  environment pre_seg fol_seg
#> 1      RELIES    AY   PRICE     PRIZE    elsewhere       L       Z
#> 2        TIME    AY   PRICE     PRIZE    elsewhere       T       M
#> 3   SURVIVORS    AY   PRICE     PRIZE    elsewhere       V       V
#> 4     SCIENCE    AY   PRICE     PRIZE    elsewhere       S     AH0
#> 5        EYES    AY   PRICE     PRIZE    elsewhere       N       Z
#> 6        TIME    AY   PRICE     PRIZE    elsewhere       T       M
#> 7  IDENTIFIED    AY   PRICE     PRIZE    elsewhere       F       D
#> 8       RIGHT    AY   PRICE     PRICE prevoiceless       R       T
#> 9    BRIGHTLY    AY   PRICE     PRICE prevoiceless       R       T
#> 10    LIBRARY    AY   PRICE     PRIZE    elsewhere       L       B
#> 11   INSIDERS    AY   PRICE     PRIZE    elsewhere       S       D
#> 12       TIME    AY   PRICE     PRIZE    elsewhere       T       M
#> 13       HIGH    AY   PRICE     PRIZE    elsewhere      HH        
#> 14       LIKE    AY   PRICE     PRICE prevoiceless       L       K
#> 15      ITEMS    AY   PRICE     PRICE prevoiceless               T
#> 16       SIZE    AY   PRICE     PRIZE    elsewhere       S       Z
#> 17      MILES    AY   PRICE     PRIZE    elsewhere       M       L
#> 18     TRIBES    AY   PRICE     PRIZE    elsewhere       R       B
#> 19  RECOGNIZE    AY   PRICE     PRIZE    elsewhere       N       Z
#> 20      TIMER    AY   PRICE     PRIZE    elsewhere       T       M

# Collapsing distinctions can be done post hoc (though it may take extra work to get the environment column to match.)
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  # Get a subset for demonstration purposes
  filter(allophone %in% c("BIT", "BIG")) %>%
  group_by(allophone) %>%
  slice_sample(n = 5) %>%
  ungroup() %>%
  # Now collapse distinctions
  mutate(allophone = fct_collapse(allophone, "BIT" = c("BIT", "BIG")),
         environment = ifelse(allophone == "BIT", "elsewhere", allophone))
#> # A tibble: 10 × 7
#>    word      vowel phoneme allophone environment pre_seg fol_seg
#>    <chr>     <chr> <fct>   <fct>     <chr>       <chr>   <chr>  
#>  1 EXISTENCE IH    KIT     BIT       elsewhere   IY0     G      
#>  2 BEGINNING IH    KIT     BIT       elsewhere   B       G      
#>  3 EXISTED   IH    KIT     BIT       elsewhere   Z       G      
#>  4 BIG       IH    KIT     BIT       elsewhere   B       G      
#>  5 EXOTIC    IH    KIT     BIT       elsewhere   V       G      
#>  6 GIVE      IH    KIT     BIT       elsewhere   G       V      
#>  7 WITH      IH    KIT     BIT       elsewhere   W       TH     
#>  8 MISSION   IH    KIT     BIT       elsewhere   M       SH     
#>  9 HIS       IH    KIT     BIT       elsewhere   HH      Z      
#> 10 EXISTED   IH    KIT     BIT       elsewhere   Z       S      

# Creating new allophones depends on the complexity of the allophone         
darla %>%
  code_allophones(.old_col = phoneme, 
                  .new_cols = c("allophone", "environment"), 
                  .fol_seg = fol_seg, 
                  .pre_seg = pre_seg) %>%
  # Create voice and voiceless distinctions for MOUTH
  mutate(allophone = case_when(phoneme == "MOUTH" & fol_seg %in% c("P", "T", "K", "CH", "F", "TH", "S", "SH") ~ "BOUT",
                               phoneme == "MOUTH" ~ "LOUD",
                               TRUE ~ allophone),
         environment = if_else(allophone == "BOUT",  "prevoiceless", environment)) %>%
  # Get a subset for demonstration purposes
  filter(phoneme == "MOUTH") %>%
  group_by(allophone) %>%
  slice_sample(n = 5) %>%
  ungroup()
#> # A tibble: 10 × 7
#>    word      vowel phoneme allophone environment  pre_seg fol_seg
#>    <chr>     <chr> <fct>   <chr>     <chr>        <chr>   <chr>  
#>  1 OUT       AW    MOUTH   BOUT      prevoiceless T       T      
#>  2 OUT       AW    MOUTH   BOUT      prevoiceless L       T      
#>  3 OUT       AW    MOUTH   BOUT      prevoiceless T       T      
#>  4 OUT       AW    MOUTH   BOUT      prevoiceless D       T      
#>  5 OUT       AW    MOUTH   BOUT      prevoiceless M       T      
#>  6 POWERLESS AW    MOUTH   LOUD      elsewhere    P       ER0    
#>  7 GROUND    AW    MOUTH   LOUD      elsewhere    R       N      
#>  8 DOWN      AW    MOUTH   LOUD      elsewhere    D       N      
#>  9 HOW       AW    MOUTH   LOUD      elsewhere    HH      HH     
#> 10 COUNTER   AW    MOUTH   LOUD      elsewhere    K       N