Code allophones

A function to classify vowel data into contextual allophones.

code_allophones(
  .df,
  .old_col,
  .new_cols = c("allophone", "allophone_environment"),
  .pre_seg,
  .fol_seg,
  .coronals = c("T", "D", "S", "Z", "SH", "ZH", "JH", "N"),
  .voiceless = c("P", "T", "K", "CH", "F", "TH", "S", "SH")
)

Arguments

.df

The dataset containing vowel data.

.old_col

The unquoted name of the column containing the vowel labels. Often called "vowel" or "phoneme" in many datasets. Note that the function assumes Wells lexical sets (FLEECE, TRAP, etc.) rather than ARPABET (IY, AE, etc.) or IPA (i, æ, etc.). If your vowels are not already coded using Wells' labels you can quickly do so with switch_transcriptions or one of the shortcuts like arpa_to_wells

.new_cols

A vector of two strings containing the names of the columns you would like to use. By default c("allophone", "allophone_environment"). The first name becomes the name of the column containing the new allophone labels. The second column becomes the name of the column describing those labels.

.pre_seg

The unquoted name of the column that contains the labels for the previous segement. In DARLA-generated spreadsheets, this is `pre_seg` and in FastTrack-generated spreadsheets, it's `previous_sound`. Assumes ARPABET labels.

.fol_seg

The unquoted name of the column that contains the labels for the following segement. In DARLA-generated spreadsheets, this is `fol_seg` and in FastTrack-generated spreadsheets, it's `next_sound`. Assumes ARPABET labels.

.coronals

A vector of strings containing ARPABET labels for coronal consonants. By default, c("T", "D", "S", "Z", "SH", "ZH", "JH", "N"). This is used to create the `TOOT` allophone of `GOOSE`.

.voiceless

A vector of strings containing ARPABET labels for voiceless consonants. By default, c("P", "T", "K", "CH", "F", "TH", "S", "SH"). This is used to create the `PRICE` allophone of `PRIZE`.

Value

A dataframe with two additional columns. One column contains labels for the allophones and the other contains category labels for those allophones' contexts. The second column can be useful for quickly excluding certain allophones like prelaterals or prerhotics or coloring families of allophones in visualizations (such as turning all prelateral allophones gray). These two new columns are positioned immediately after the original vowel column indicated in .old_col,

Note

Here are the list of the contextual allophones that are created. Note that I largely follow my own advice about what to call elsewhere allophones, what to call prelateral allophones, and other allophones. Obviously, this list is pretty subjective and largely based on what my own research has needed, so it may not work completely for you and your research. Please contact me at joey_stanley@byu.edu if you want to see an allophone get added or if you spot an error in the coding.

FLEECE becomes

ZEAL before laterals

BEET elsewhere

KIT becomes

GUILT before laterals

NEAR before rhotics

BIG before G

BIN before M and N

BING before NG

BIT elsewhere

FACE becomes

FLAIL before laterals

VAGUE before G

BAIT elsewhere

DRESS becomes

SHELF before laterals

SQUARE before rhotics

BEG before G

BEN before M and N

BENG before NG

BET elsewhere

TRAP becomes

TALC before laterals

BAG before G

BAN before M and N

BANG before NG

BAT elsewhere

LOT becomes

GOLF before laterals

START before rhotics

BOT elsewhere

THOUGHT becomes

FAULT before laterals

FORCE befpre rhotics

BOUGHT elsewhere

STRUT becomes

MULCH before laterals

BUT elsewhere

GOAT becomes

JOLT before laterals

BOAT elsewhere

FOOT becomes

WOLF before laterals

CURE before rhotics

PUT elsewhere

GOOSE becomes

MULE before Y

TOOT before coronals

SPOOL before laterals

BOOT elsewhere

PRICE becomes

PRICE before voiceless segments

PRIZE elsewhere

Unfortunately, it is not straightforward to customize this list but you can always copy the source code and modify the list yourself.

Alternatively, you can use forcats::fct_collapse() to collapse distinctions that you don't need. See example code below.

You can also of course create your own allophones if desired. Note that some allophones depend on other environmental information like syllable structure and morpheme/word boundaries, or they may be entirely lexical (FORCE vs. NORTH). They may be more complicated than what ARPABET can code for (MARY, MERRY, and MARRY) or just inconsistently coded. For the sake of simplicity, these allophones are not included in this function.

The environments therefore are the following

"prelateral" includes ZEAL, GUILT, FLAIL, SHELF, TALC, GOLF, FAULT, MULCH, JOLT, WOLF, SPOOL

"prerhotic" includes NEAR, SQUARE, START, FORCE, CURE

"prevelar" includes BIG, VAGUE, BEG, BAG,

"prenasal" includes BIN, BEN, BAN

"prevelarnasal" includes BING, BENG, BANG

"prevoiceless" includes PRICE

"post-Y" includes MULE

"postcoronal" includes TOOT

"elsewhere" includes BEET, BIT, BAIT, BET, BAT, BOT, BOUGHT, BUT, BOAT, PUT, BOOT, PRIZE

Examples

suppressPackageStartupMessages(library(tidyverse))
#> Warning: package ‘ggplot2’ was built under R version 4.1.2
#> Warning: package ‘tibble’ was built under R version 4.1.2
#> Warning: package ‘tidyr’ was built under R version 4.1.2
#> Warning: package ‘readr’ was built under R version 4.1.2
#> Warning: package ‘stringr’ was built under R version 4.1.2

# Get some sample DARLA data to play with
darla <- joeysvowels::darla %>%
  select(word, vowel, pre_seg, fol_seg) %>%
  mutate(phoneme = joeyr:::arpa_to_wells(vowel), .after = vowel)

# Basic usage
darla %>%
  code_allophones(.old_col = phoneme, .fol_seg = fol_seg, .pre_seg = pre_seg) %>%
  slice_sample(n = 20)
#>            word vowel phoneme allophone allophone_environment pre_seg fol_seg
#> 1     DESTROYED    OY  CHOICE    CHOICE             elsewhere       R       D
#> 2          ALSO    OW    GOAT      BOAT             elsewhere       S       P
#> 3       ANYBODY    EH   DRESS       BEN              prenasal       D       N
#> 4       PARTIES    IY  FLEECE      BEET             elsewhere       T       Z
#> 5         PAPER    ER   NURSE     NURSE             elsewhere       P     IH1
#> 6            ON    AO THOUGHT    BOUGHT             elsewhere      CH       N
#> 7       LIBRARY    IY  FLEECE      BEET             elsewhere       R        
#> 8          TOOL    UW   GOOSE      TOOT           postcoronal       T       L
#> 9  MIRACULOUSLY    ER   NURSE     NURSE             elsewhere       M     AE1
#> 10 PERFORMANCES    AO THOUGHT     FORCE             prerhotic       F       R
#> 11     ALTHOUGH    AO THOUGHT     FAULT            prelateral       D       L
#> 12          BOX    AA     LOT       BOT             elsewhere       B       K
#> 13       POLICE    IY  FLEECE      BEET             elsewhere       L       S
#> 14      NOTHING    AH   STRUT       BUT             elsewhere       N      TH
#> 15      SETTING    IH     KIT      BING         prevelarnasal       T      NG
#> 16        NEVER    ER   NURSE     NURSE             elsewhere       V       F
#> 17           AS    EH   DRESS       BET             elsewhere               Z
#> 18      HOSTESS    OW    GOAT      BOAT             elsewhere      HH       S
#> 19        LOOPS    UW   GOOSE      BOOT             elsewhere       L       P
#> 20       PEOPLE    AH   STRUT     MULCH            prelateral       P       L

# Specify the names of the new columns with the `.new_cols` argument
darla %>%
  code_allophones(.old_col = phoneme,
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg,
                  .pre_seg = pre_seg) %>%
  slice_sample(n = 20)
#>          word vowel phoneme allophone  environment pre_seg fol_seg
#> 1  ALTOGETHER    AO THOUGHT     FAULT   prelateral       L       L
#> 2      BRIGHT    AY   PRICE     PRICE prevoiceless       R       T
#> 3       FORTY    AO THOUGHT     FORCE    prerhotic       F       R
#> 4      CUTTER    AH   STRUT       BUT    elsewhere       K       T
#> 5       DOING    UW   GOOSE      TOOT  postcoronal       D     IH0
#> 6       GROOM    UW   GOOSE      BOOT    elsewhere       R       M
#> 7          ET    EH   DRESS       BET    elsewhere               T
#> 8      COLONY    IY  FLEECE      BEET    elsewhere       N       K
#> 9     VIEWING    UW   GOOSE      MULE       post-Y       Y     IH0
#> 10     CHANGE    EY    FACE      BAIT    elsewhere      CH       N
#> 11  BATTERIES    ER   NURSE     NURSE    elsewhere       T     IY0
#> 12      HOUSE    AW   MOUTH     MOUTH    elsewhere      HH       S
#> 13       GAIL    EY    FACE     FLAIL   prelateral       G       L
#> 14       TEXT    EH   DRESS       BET    elsewhere       T       K
#> 15        ANY    EH   DRESS       BEN     prenasal       N       N
#> 16      THIRD    ER   NURSE     NURSE    elsewhere      TH       D
#> 17       LOGS    AO THOUGHT    BOUGHT    elsewhere       L       G
#> 18       BLUE    UW   GOOSE      BOOT    elsewhere       L       T
#> 19   PATIENTS    EY    FACE      BAIT    elsewhere       P      SH
#> 20         AS    EH   DRESS       BET    elsewhere               Z

# Filtering by environment is straightforward
darla %>%
  code_allophones(.old_col = phoneme,
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg,
                  .pre_seg = pre_seg) %>%
  filter(environment == "elsewhere") %>%
  slice_sample(n = 20)
#>        word vowel phoneme allophone environment pre_seg fol_seg
#> 1      FIND    AY   PRICE     PRIZE   elsewhere       F       N
#> 2      LOOK    UH    FOOT       PUT   elsewhere       L       K
#> 3        AT    AE    TRAP       BAT   elsewhere       T       T
#> 4       NOW    AW   MOUTH     MOUTH   elsewhere       N      DH
#> 5  CROCKETT    IH     KIT       BIT   elsewhere       K       T
#> 6   DOESN'T    AH   STRUT       BUT   elsewhere       D       Z
#> 7  SUPPOSED    OW    GOAT      BOAT   elsewhere       P       Z
#> 8      SLUT    AH   STRUT       BUT   elsewhere       L       T
#> 9      BOOK    UH    FOOT       PUT   elsewhere       B       K
#> 10     MANY    IY  FLEECE      BEET   elsewhere       N       F
#> 11    OZONE    OW    GOAT      BOAT   elsewhere       K       Z
#> 12 THOUSAND    AW   MOUTH     MOUTH   elsewhere      TH       Z
#> 13     HEAD    EH   DRESS       BET   elsewhere      HH       D
#> 14  THOUGHT    AO THOUGHT    BOUGHT   elsewhere      TH       T
#> 15     TIME    AY   PRICE     PRIZE   elsewhere       T       M
#> 16      SEE    IY  FLEECE      BEET   elsewhere       S       W
#> 17  WITHOUT    AW   MOUTH     MOUTH   elsewhere      TH       T
#> 18    MONEY    IY  FLEECE      BEET   elsewhere       N     IH1
#> 19  RIBBONS    AH   STRUT       BUT   elsewhere       B       N
#> 20     LONG    AO THOUGHT    BOUGHT   elsewhere       L      NG
darla %>%
  code_allophones(.old_col = phoneme,
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg,
                  .pre_seg = pre_seg) %>%
  filter(!environment %in% c("prerhotic", "prevelarnasal", "prevelar")) %>%
  slice_sample(n = 20)
#>              word vowel phoneme allophone  environment pre_seg fol_seg
#> 1         FURTHER    ER   NURSE     NURSE    elsewhere       F      DH
#> 2            MOVE    UW   GOOSE      BOOT    elsewhere       M       V
#> 3         SYSTEMS    AH   STRUT       BUT    elsewhere       T       M
#> 4             HER    ER   NURSE     NURSE    elsewhere      HH       L
#> 5           BEGAN    AE    TRAP       BAN     prenasal       G       N
#> 6           DOING    UW   GOOSE      TOOT  postcoronal       D     IH0
#> 7              OR    ER   NURSE     NURSE    elsewhere     UW1      TH
#> 8             CHE    EY    FACE      BAIT    elsewhere      CH     IH1
#> 9    INTELLECTUAL    IH     KIT       BIN     prenasal       N       N
#> 10        BEDROOM    EH   DRESS       BET    elsewhere       B       D
#> 11         CORNER    ER   NURSE     NURSE    elsewhere       N     AH0
#> 12           LOSE    UW   GOOSE      BOOT    elsewhere       L       Z
#> 13             NO    OW    GOAT      BOAT    elsewhere       N        
#> 14           BODY    AA     LOT       BOT    elsewhere       B       D
#> 15           EVER    EH   DRESS       BET    elsewhere       Z       V
#> 16        BEHAVES    IH     KIT       BIT    elsewhere       B      HH
#> 17          THROW    OW    GOAT      BOAT    elsewhere       R       W
#> 18        GEORGIA    AH   STRUT       BUT    elsewhere      JH        
#> 19 ADMINISTRATORS    EY    FACE      BAIT    elsewhere       R       T
#> 20          SITES    AY   PRICE     PRICE prevoiceless       S       T

# Some users may want to supply their own list of coronal consonants.
darla %>%
  code_allophones(.old_col = phoneme,
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg,
                  .pre_seg = pre_seg,
                  .coronals = c("T", "D", "S", "Z", "SH", "ZH", "JH", "N", "Y")) %>%
  filter(phoneme == "GOOSE") %>%
  slice_sample(n = 20)
#>        word vowel phoneme allophone environment pre_seg fol_seg
#> 1        TO    UW   GOOSE      TOOT postcoronal       T       L
#> 2     STOOL    UW   GOOSE      TOOT postcoronal       T       L
#> 3      FOOL    UW   GOOSE     SPOOL  prelateral       F       L
#> 4        TO    UW   GOOSE      TOOT postcoronal       T     SIL
#> 5      COOL    UW   GOOSE     SPOOL  prelateral       K       L
#> 6      TOOL    UW   GOOSE      TOOT postcoronal       T       L
#> 7  BALLOONS    UW   GOOSE      BOOT   elsewhere       L       N
#> 8        TO    UW   GOOSE      TOOT postcoronal       T       T
#> 9     GOULD    UW   GOOSE     SPOOL  prelateral       G       L
#> 10      WHO    UW   GOOSE      BOOT   elsewhere      HH        
#> 11      NEW    UW   GOOSE      MULE      post-Y       Y       P
#> 12 STUDENTS    UW   GOOSE      TOOT postcoronal       T       D
#> 13       TO    UW   GOOSE      TOOT postcoronal       T       P
#> 14      USE    UW   GOOSE      MULE      post-Y       Y       Z
#> 15      WHO    UW   GOOSE      BOOT   elsewhere      HH       W
#> 16  FORTUNE    UW   GOOSE      BOOT   elsewhere      CH       N
#> 17     INTO    UW   GOOSE      TOOT postcoronal       T     SIL
#> 18       TO    UW   GOOSE      TOOT postcoronal       T        
#> 19       TO    UW   GOOSE      TOOT postcoronal       T     IH0
#> 20 TROOPERS    UW   GOOSE      BOOT   elsewhere       R       P

# Other users may want to specify their own list of voiceless consonants.
darla %>%
  code_allophones(.old_col = phoneme,
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg,
                  .pre_seg = pre_seg,
                  .voiceless = c("P", "T", "K", "CH", "F", "TH", "S", "SH", "X")) %>%
  filter(phoneme == "PRICE") %>%
  slice_sample(n = 20)
#>          word vowel phoneme allophone  environment pre_seg fol_seg
#> 1        WIRE    AY   PRICE     PRIZE    elsewhere       W       R
#> 2       WHILE    AY   PRICE     PRIZE    elsewhere       W       L
#> 3       RIGHT    AY   PRICE     PRICE prevoiceless       R       T
#> 4        HIND    AY   PRICE     PRIZE    elsewhere      HH       N
#> 5      PILING    AY   PRICE     PRIZE    elsewhere       P       L
#> 6       APPLY    AY   PRICE     PRIZE    elsewhere       L      DH
#> 7      BLINDS    AY   PRICE     PRIZE    elsewhere       L       N
#> 8     CHINESE    AY   PRICE     PRIZE    elsewhere      CH       N
#> 9        EYES    AY   PRICE     PRIZE    elsewhere       N       Z
#> 10     FLIGHT    AY   PRICE     PRICE prevoiceless       L       T
#> 11      RIGHT    AY   PRICE     PRICE prevoiceless       R       T
#> 12       DIET    AY   PRICE     PRIZE    elsewhere       D     AH0
#> 13       SIDE    AY   PRICE     PRIZE    elsewhere       S       D
#> 14      RIGHT    AY   PRICE     PRICE prevoiceless       R       T
#> 15 IDENTIFIED    AY   PRICE     PRIZE    elsewhere       F       D
#> 16      DRIVE    AY   PRICE     PRIZE    elsewhere       R       V
#> 17         BY    AY   PRICE     PRIZE    elsewhere       B     EY1
#> 18     MYSELF    AY   PRICE     PRICE prevoiceless       M       S
#> 19      RIGHT    AY   PRICE     PRICE prevoiceless       R       T
#> 20    ARRIVED    AY   PRICE     PRIZE    elsewhere     ER0       V

# Collapsing distinctions can be done post hoc (though it may take extra work to get the environment column to match.)
darla %>%
  code_allophones(.old_col = phoneme,
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg,
                  .pre_seg = pre_seg) %>%
  # Get a subset for demonstration purposes
  filter(allophone %in% c("BIT", "BIG")) %>%
  group_by(allophone) %>%
  slice_sample(n = 5) %>%
  ungroup() %>%
  # Now collapse distinctions
  mutate(allophone = fct_collapse(allophone, "BIT" = c("BIT", "BIG")),
         environment = ifelse(allophone == "BIT", "elsewhere", allophone))
#> # A tibble: 10 × 7
#>    word      vowel phoneme allophone environment pre_seg fol_seg
#>    <chr>     <chr> <fct>   <fct>     <chr>       <chr>   <chr>  
#>  1 BEGINNING IH    KIT     BIT       elsewhere   B       G      
#>  2 TRIGGER   IH    KIT     BIT       elsewhere   R       G      
#>  3 EXISTED   IH    KIT     BIT       elsewhere   Z       G      
#>  4 BIG       IH    KIT     BIT       elsewhere   B       G      
#>  5 EXOTIC    IH    KIT     BIT       elsewhere   V       G      
#>  6 HARNESS   IH    KIT     BIT       elsewhere   N       S      
#>  7 GENETIC   IH    KIT     BIT       elsewhere   T       K      
#>  8 MUSIC     IH    KIT     BIT       elsewhere   Z       K      
#>  9 PLEASES   IH    KIT     BIT       elsewhere   Z       Z      
#> 10 SICK      IH    KIT     BIT       elsewhere   S       K      

# Creating new allophones depends on the complexity of the allophone
darla %>%
  code_allophones(.old_col = phoneme,
                  .new_cols = c("allophone", "environment"),
                  .fol_seg = fol_seg,
                  .pre_seg = pre_seg) %>%
  # Create voice and voiceless distinctions for MOUTH
  mutate(allophone = case_when(phoneme == "MOUTH" & fol_seg %in% c("P", "T", "K", "CH", "F", "TH", "S", "SH") ~ "BOUT",
                               phoneme == "MOUTH" ~ "LOUD",
                               TRUE ~ allophone),
         environment = if_else(allophone == "BOUT",  "prevoiceless", environment)) %>%
  # Get a subset for demonstration purposes
  filter(phoneme == "MOUTH") %>%
  group_by(allophone) %>%
  slice_sample(n = 5) %>%
  ungroup()
#> # A tibble: 10 × 7
#>    word   vowel phoneme allophone environment  pre_seg fol_seg
#>    <chr>  <chr> <fct>   <chr>     <chr>        <chr>   <chr>  
#>  1 ABOUT  AW    MOUTH   BOUT      prevoiceless B       "T"    
#>  2 HOUSE  AW    MOUTH   BOUT      prevoiceless HH      "S"    
#>  3 OUT    AW    MOUTH   BOUT      prevoiceless T       "T"    
#>  4 HOUSE  AW    MOUTH   BOUT      prevoiceless HH      "S"    
#>  5 ABOUT  AW    MOUTH   BOUT      prevoiceless B       "T"    
#>  6 NOW    AW    MOUTH   LOUD      elsewhere    N       ""     
#>  7 DOWN   AW    MOUTH   LOUD      elsewhere    D       "N"    
#>  8 HOW    AW    MOUTH   LOUD      elsewhere    HH      "HH"   
#>  9 HOUSED AW    MOUTH   LOUD      elsewhere    HH      "Z"    
#> 10 OUR    AW    MOUTH   LOUD      elsewhere    M       "R"