ANAE Vowel Normalization

This is a a tidyverse-compatible function that makes it easy to normalize your data using the method described in the Atlas of North American English. NOTE: This is brand new and has not been tested robustly.

norm_anae(df, hz_cols, vowel_id, speaker_id, use_telsur_g = TRUE)

Arguments

df
The dataframe containing the formant measurements you want to normalize.

hz_cols
A list of columns (unquoted) containing the formant measurements themselves.

vowel_id
The name of the column containing unique identifiers per vowel token. If your data is set up so that there is one row per token, you can put row.names(.) here instead.

speaker_id
The name of the column containing unique identifiers per speaker (usually the column containing the speaker name).

use_telsur_g
By default, this will use the Telsur G value (6.896874) listed in the ANAE. If set to FALSE, it will calculate the G value based on the dataset.

Value

The same dataframe, but with new column(s), suffixed with "_anae" that have the normalized data.

Details

The data must be grouped by speaker prior to running the function.

The function works best when only F1 and F2 data are included. F3 can be included but the results may not be comparable with other studies.

By default, the function will use the Telsur G value listed in the ANAE (6.896874). This is recommended to make results most compatible with the ANAE and other studies that use the same normalization procedure. The function can calculate a G value based on the dataset provided when use_telsur_g is set to FALSE.

It is unclear how the ANAE function should work with trajectory data. This function pools all data together and normalizes it together, which means one small modification was required to calculate the G value if the Telsur G is not used: I had to add the average number of time points per vowel token in the denominator. Not sure if that's how it should be done, but it makes sense to me and returns sensible results.

Examples

library(tidyverse)

#> Error in library(tidyverse): there is no package called ‘tidyverse’

df <- joeysvowels::idahoans

#> Error in loadNamespace(name): there is no package called ‘joeysvowels’

# Run the function. Since this data doesn't have a column for token ids, I'll create one here.
df %>%
   group_by(speaker) %>%
   norm_anae(hz_cols = c(F1, F2), vowel_id = row.names(.), speaker_id = speaker) %>%
   ungroup() %>%
   select(F1, F2, F1_anae, F2_anae) # <- just the relevant columns

#> Error in UseMethod("group_by"): no applicable method for 'group_by' applied to an object of class "function"