Calculate Mahalanobis Distance

This is a tidyverse-tidyverse-compatible version of the `mahalanobis` function. It just makes it easier to include it as part of a `dplyr::mutate`. One small modification that `mahalnobis` function does not do is that if there are fewer than 5 measurements in the dataset, , it returns them all as having a distance of zero. Prevents some errors that way. Requires the MASS package to be installed, but this this not necessarily loaded.

tidy_mahalanobis(...)

Arguments

...
Names of columns that should be included in the Mahalanobis distance. For vowel data, this is typically your F1 and F2 columns.

Value

A vector that contains the Mahalanobis distances for each observation.

Examples

require(dplyr)

# Create some artificial vowel data
tibble(F1 = rnorm(50, 0, 1),
       F2 = rnorm(50, 0, 1)) %>%
 # Calculate the Mahalanobis distance
 mutate(mahal_dist = tidy_mahalanobis(F1, F2))

#> # A tibble: 50 x 3
#>         F1       F2 mahal_dist
#>      <dbl>    <dbl>      <dbl>
#>  1  0.110  -0.685        0.366
#>  2  0.853   0.337        0.825
#>  3 -0.234  -0.412        0.284
#>  4  2.09    0.934        6.08 
#>  5 -0.111   1.84         4.94 
#>  6 -1.39   -0.705        3.78 
#>  7 -1.14    0.00851      2.90 
#>  8  1.70    2.03         7.20 
#>  9 -0.0801 -1.34         1.67 
#> 10 -0.437   1.16         2.97 
#> # … with 40 more rows

# Same thing, but using base R: 
vowel_data <- data.frame(F1 = rnorm(50, 0, 1),
                         F2 = rnorm(50, 0, 1)) 
vowel_data$mahal_dist <- tidy_mahalanobis(vowel_data$F1, vowel_data$F2)
head(vowel_data)

#>           F1          F2 mahal_dist
#> 1  0.1153482  1.89631708   5.664362
#> 2 -1.7680484 -2.06007072   8.848384
#> 3 -1.4076389  0.06454387   1.825010
#> 4  0.7091785 -0.26514740   1.084924
#> 5 -1.2408429 -0.44734453   1.552218
#> 6 -0.3683273 -1.41070093   2.783126